Move file nesting to combine all V2 in one folder
